requirements.txt
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.5
openai==1.6.1
python-dotenv==1.0.0
pandas==2.1.4
tenacity==8.2.3
streamlit==1.29.0
plotly==5.18.0
